requirements.txt
azure-functions==1.21.3
azure-data-tables==12.6.0
azure-identity==1.25.2
